Implement CommunicationInterface to replace MQTTClient and streamline MQTT handling
